+++ derive-builder-test/tests/owned.rs
@@ -1,7 +1,7 @@
 #[macro_use] extern crate derive_builder;
 
 #[derive(Debug, PartialEq, Default, Builder, Clone)]
-#[owned]
+#[setters(owned)]
 struct Lorem {
     ipsum: String,
 }
@@ -22,11 +22,12 @@ fn contructor_sanity_check() {
 }
 
 #[test]
-fn consuming() {
-    let f: fn(Lorem, String) -> Lorem = Lorem::ipsum;
+fn consuming_setter() {
+    // the setter must have the correct signature
+    let consuming_setter: fn(Lorem, String) -> Lorem = Lorem::ipsum;
 
     let old = Lorem::new("lorem");
-    let new = f(old, "new".to_string());
+    let new = consuming_setter(old, "new".to_string());
 
     assert_eq!(new, Lorem { ipsum: "new".into() });
 }